added password check for private lobbies, added play/id need fix for null
