+++ my-app/.env
@@ -3,3 +3,5 @@ NEXTAUTH_SECRET="ud7HdzRf5TwFCEa167R59PaUkV0XnhTu37DJnXmis48="
 
 GITHUB_ID="2e2cbf0a3f094d1e998c"
 GITHUB_SECRET="611d491746da96d20793695d97bf7ceea271f8b8"
+
+NEXT_PUBLIC_CONVEX_URL="https://trustworthy-meadowlark-912.convex.cloud"
