+++ my-app/app/not-found.tsx
@@ -1,0 +1,10 @@
+
+
+export default function notFound() {
+    return (
+        <div>
+        <h1>404</h1>
+        </div>
+    );
+    }
+
